Revert "basically finish drawing the ui of login screen"
