+++ .idea/deploymentTargetDropDown.xml
@@ -2,7 +2,10 @@
 <project version="4">
   <component name="deploymentTargetDropDown">
     <value>
-      <entry key="CustomInputBoxPreview">
+      <entry key="PreviewSignUpScreen">
+        <State />
+      </entry>
+      <entry key="SplashScreenPreview">
         <State />
       </entry>
       <entry key="app">
