--- app/src/main/java/com/example/meditation/ui/screen/MineScreen.kt
@@ -1,8 +1,0 @@
-package com.example.meditation.ui.screen
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun MineScreen(){
-
-}
